% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpdfs.R
\name{plot_pdf_by_stat}
\alias{plot_pdf_by_stat}
\title{Plot PDF with statistical value.}
\usage{
plot_pdf_by_stat(
  distribution = "norm",
  xrange = c(-3, 3),
  stat_value,
  side = "lower",
  line_color = "red",
  area_color = "deeppink2",
  ...
)
}
\arguments{
\item{distribution}{Character string indicating probability distribution function. Eg "norm", "t", "chisq". Internally, "d-", "p-" are added as prefix of the string. eg. dnorm(), pnorm().}

\item{xrange}{X axis range for plot.}

\item{stat_value}{Statistical value to show on graph.}

\item{side}{Either, "lower" "upper" "two-side-inner" or "two-side-outer"}

\item{line_color}{color for vertical line at stat_value.}

\item{area_color}{color for fill aesthetics in graph.}

\item{...}{Arguments passed on to distribution. eg mean=1, sd = 2 for "norm".}
}
\description{
This function will plot PDF with statistical value.
}
\examples{
plot_pdf_by_stat(stat_value = 1.4,xrange = c(-3,3), side = "lower")
}
